fix: no data message moved on smart data view module
