+++ apps/gauzy/src/app/pages/recurring-expense-employee/recurring-expense-employee.module.ts
@@ -12,12 +12,7 @@ import {
 } from '@nebular/theme';
 import { NgxPermissionsModule } from 'ngx-permissions';
 import { TranslateModule } from '@ngx-translate/core';
-import {
-	GauzyButtonActionModule,
-	NoDataMessageModule,
-	RecurringExpenseBlockModule,
-	SharedModule
-} from '@gauzy/ui-core/shared';
+import { RecurringExpenseBlockModule, SmartDataViewLayoutModule, SharedModule } from '@gauzy/ui-core/shared';
 import { RecurringExpensesEmployeeRoutingModule } from './recurring-expense-employee-routing.module';
 import { RecurringExpensesEmployeeComponent } from './recurring-expense-employee.component';
 
@@ -38,8 +33,7 @@ import { RecurringExpensesEmployeeComponent } from './recurring-expense-employee
 		SharedModule,
 		RecurringExpensesEmployeeRoutingModule,
 		RecurringExpenseBlockModule,
-		GauzyButtonActionModule,
-		NoDataMessageModule
+		SmartDataViewLayoutModule
 	],
 	declarations: [RecurringExpensesEmployeeComponent]
 })
