--- packages/ui-core/shared/src/lib/no-data-message/index.ts
@@ -1,2 +1,0 @@
-export * from './no-data-message.component';
-export * from './no-data-message.module';
