+++ packages/ui-core/shared/src/lib/report/activities-report-grid/activities-report-grid.module.ts
@@ -4,7 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { NbCardModule, NbIconModule, NbSelectModule, NbSpinnerModule } from '@nebular/theme';
 import { TranslateModule } from '@ngx-translate/core';
 import { SharedModule } from '../../shared.module';
-import { NoDataMessageModule } from '../../no-data-message/no-data-message.module';
+import { NoDataMessageModule } from '../../smart-data-layout/no-data-message/no-data-message.module';
 import { ActivitiesReportGridComponent } from './activities-report-grid.component';
 import { ProgressStatusModule } from '../../progress-status/progress-status.module';
 import { ProjectColumnViewModule } from '../project-column-view/project-column-view.module';
