+++ packages/ui-core/shared/src/lib/report/amounts-owed-grid/amounts-owed-grid.module.ts
@@ -4,7 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { NbCardModule, NbIconModule, NbSelectModule, NbSpinnerModule } from '@nebular/theme';
 import { TranslateModule } from '@ngx-translate/core';
 import { SharedModule } from '../../shared.module';
-import { NoDataMessageModule } from '../../no-data-message/no-data-message.module';
+import { NoDataMessageModule } from '../../smart-data-layout/no-data-message/no-data-message.module';
 import { AmountsOwedGridComponent } from './amounts-owed-grid.component';
 
 @NgModule({
