+++ packages/ui-core/shared/src/lib/report/daily-grid/daily-grid.module.ts
@@ -14,7 +14,7 @@ import { SharedModule } from '../../shared.module';
 import { TableComponentsModule } from '../../table-components/table-components.module';
 import { DailyGridComponent } from './daily-grid.component';
 import { ProjectColumnViewModule } from '../project-column-view/project-column-view.module';
-import { NoDataMessageModule } from '../../no-data-message/no-data-message.module';
+import { NoDataMessageModule } from '../../smart-data-layout/no-data-message/no-data-message.module';
 
 @NgModule({
 	imports: [
