+++ packages/ui-core/shared/src/lib/smart-data-layout/index.ts
@@ -1,7 +1,9 @@
 // Modules
 export * from './smart-data-view-layout.module';
+export * from './no-data-message/no-data-message.module';
 
 // Components
+export * from './no-data-message/no-data-message.component';
 export * from './smart-table-toggle/smart-table-toggle.component';
 export * from './pagination/pagination.component';
 export * from './pagination/pagination-filter-base.component';
