+++ packages/ui-core/shared/src/lib/smart-data-layout/no-data-message/no-data-message.component.scss
