+++ packages/ui-core/shared/src/lib/smart-data-layout/smart-data-view-layout.module.ts
@@ -5,10 +5,11 @@ import { TranslateModule } from '@ngx-translate/core';
 import { Angular2SmartTableModule } from 'angular2-smart-table';
 import { NgxPermissionsModule } from 'ngx-permissions';
 import { SharedModule } from '../shared.module';
+import { GauzyButtonActionModule } from '../gauzy-button-action/gauzy-button-action.module';
+import { NoDataMessageModule } from './no-data-message/no-data-message.module';
 import { PaginationComponent } from './pagination/pagination.component';
 import { PaginationV2Component } from './pagination/pagination-v2/pagination-v2.component';
 import { SmartTableToggleComponent } from './smart-table-toggle/smart-table-toggle.component';
-import { GauzyButtonActionModule } from '../gauzy-button-action';
 
 // Nebular Modules
 const NB_MODULES = [NbToggleModule, NbIconModule, NbSelectModule];
@@ -25,9 +26,10 @@ const COMPONENTS = [PaginationComponent, PaginationV2Component, SmartTableToggle
 		NgxPermissionsModule.forChild(),
 		Angular2SmartTableModule,
 		SharedModule,
-		GauzyButtonActionModule
+		GauzyButtonActionModule,
+		NoDataMessageModule
 	],
-	exports: [Angular2SmartTableModule, GauzyButtonActionModule, ...COMPONENTS],
+	exports: [Angular2SmartTableModule, GauzyButtonActionModule, NoDataMessageModule, ...COMPONENTS],
 	providers: []
 })
 export class SmartDataViewLayoutModule {}
